Move agents into main ocs library
Note that Dockerfiles for the Agents are all going to still be broken.
