+++ ocs/agents/ocs_plugin_standard.py
@@ -7,12 +7,14 @@ path and called something like ocs_plugin_*.
 import ocs
 import os
 root = os.path.abspath(os.path.split(__file__)[0])
+print(root)
 
 for n,f in [
-        ('RegistryAgent', 'registry/registry.py'),
-        ('AggregatorAgent', 'aggregator/aggregator_agent.py'),
-        ('HostManager', 'host_manager/host_manager.py'),
-        ('FakeDataAgent', 'fake_data/fake_data_agent.py'),
-        ('InfluxDBAgent', 'influxdb_publisher/influxdb_publisher.py'),
+        ('RegistryAgent', 'registry.py'),
+        ('AggregatorAgent', 'aggregator_agent.py'),
+        ('HostManager', 'host_manager.py'),
+        ('FakeDataAgent', 'fake_data_agent.py'),
+        ('InfluxDBAgent', 'influxdb_publisher.py'),
+        ('BarebonesAgent', 'barebones_agent.py'),
 ]:
     ocs.site_config.register_agent_class(n, os.path.join(root, f))
