+++ ocs/agents/registry.py
