+++ tests/agents/test_host_manager.py
@@ -1,6 +1,4 @@
-import sys
-sys.path.insert(0, '../agents/host_manager/')
-from host_manager import HostManager
+from ocs.agents.host_manager import HostManager
 
 import pytest_twisted
