+++ tests/agents/test_ocs_plugin_standard.py
@@ -1,6 +1,4 @@
-import sys
-sys.path.insert(0, '../agents/')
-import ocs_plugin_standard
+from ocs.agents import ocs_plugin_standard
 
 
 def test_agent_script_reg():
