+++ tests/integration/test_fake_data_agent_integration.py
@@ -12,11 +12,13 @@ from integration.util import (
     create_crossbar_fixture,
 )
 
+AGENT_PATH = '../ocs/agents/fake_data_agent.py'
+
 pytest_plugins = ("docker_compose")
 
 wait_for_crossbar = create_crossbar_fixture()
 run_agent = create_agent_runner_fixture(
-    '../agents/fake_data/fake_data_agent.py', 'fake_data')
+    AGENT_PATH, 'fake_data')
 client = create_client_fixture('fake-data-local')
 
 
@@ -48,7 +50,7 @@ def test_fake_data_agent_acq(wait_for_crossbar, run_agent, client):
 
 # Test autostartup
 run_agent_acq = create_agent_runner_fixture(
-    '../agents/fake_data/fake_data_agent.py',
+    AGENT_PATH,
     'fake_data',
     args=['--mode', 'acq'])
